requirements.txt
Django==5.1.7
djangorestframework
markdown
django-filter
psycopg2
